+++ velesresearch/website_template/public/reload.js
@@ -2,11 +2,11 @@ const ws = new WebSocket('ws://localhost:8080');
 
 ws.onmessage = (event) => {
     if (event.data === 'reload') {
-        console.log('Reloading page due to file change...');
-        window.location.reload();  // Reload the page when the server signals a change
+        console.log('Reloading the survey due to file change...');
+        loadjs(`main.js?t=${new Date().getTime()}`)
     }
 };
 
 ws.onopen = () => {
-    console.log('WebSocket connection established');
+    console.log('WebSocket connection for hot reloading established');
 };